refactor(cli): move resolving logic out of verify
