+++ crates/cli/src/resolver.rs
@@ -1,12 +1,13 @@
 use camino::Utf8PathBuf;
 use console::{style, Emoji};
-use indicatif::{ProgressBar, ProgressStyle};
+use indicatif::ProgressStyle;
 use serde::{Deserialize, Serialize};
 use std::{fs, time::Instant};
 use walkdir::{DirEntry, WalkDir};
 
 use crate::api::{FileInfo, ProjectMetadataInfo};
 use dyn_compiler::dyn_compiler::{DynamicCompiler, SupportedCairoVersions, SupportedScarbVersions};
+use voyager_resolver_cairo::compiler::scarb_utils::read_additional_scarb_manifest_metadata;
 use voyager_resolver_cairo::dyn_compiler::VoyagerGeneratorWrapper as VoyagerGenerator;
 
 pub fn get_dynamic_compiler(cairo_version: SupportedCairoVersions) -> Box<dyn DynamicCompiler> {
@@ -31,15 +32,15 @@ pub fn resolve_scarb(
     cairo_version: SupportedCairoVersions,
     scarb_version: SupportedScarbVersions,
 ) -> anyhow::Result<(Vec<FileInfo>, ProjectMetadataInfo)> {
-    // Start a spinner for the verification process
+    // Start a spinner for the resolving process
     let started = Instant::now();
-    // let spinner_style = ProgressStyle::with_template("{prefix:.bold.dim} {spinner} {wide_msg}")
-    //     .unwrap()
-    //     .tick_chars("⠁⠂⠄⡀⢀⠠⠐⠈");
+    let spinner_style = ProgressStyle::with_template("{prefix:.bold.dim} {spinner} {wide_msg}")
+        .unwrap()
+        .tick_chars("⠁⠂⠄⡀⢀⠠⠐⠈");
 
     println!(
         "{} {} Resolving contract: Extracting files from the Scarb project...",
-        style("[1/4]").bold().dim(),
+        style("[1/3]").bold().dim(),
         Emoji("📃  ", "")
     );
     // Extract necessary files from the Scarb project for the verified contract
@@ -52,7 +53,6 @@ pub fn resolve_scarb(
     };
 
     let compiler = get_dynamic_compiler(cairo_version);
-
     let contract_paths = compiler.get_contracts_to_verify_path(&source_dir)?;
 
     // TODO move the contract selection before the resolving step as a 'pre-resolving' step
@@ -71,19 +71,15 @@ pub fn resolve_scarb(
         style("[2/3]").bold().dim(),
         Emoji("🔗  ", "")
     );
-    let steps = 4;
-    let pb = ProgressBar::new(steps);
 
     // Read the scarb metadata to get more information
+    // TODO: switch this to using scarb-metadata
     let scarb_toml_content = fs::read_to_string(source_dir.join("Scarb.toml"))?;
-    let scarb_metadata_package_name = toml::from_str::<ScarbTomlRawData>(&scarb_toml_content)?
-        .package
-        .name;
-    pb.inc(1);
+    let extracted_scarb_toml_data =
+        read_additional_scarb_manifest_metadata(scarb_toml_content.as_str())?;
 
     // Compiler and extract the necessary files
     compiler.compile_project(&source_dir)?;
-    pb.inc(2);
 
     // Since we know that we extract the files into the `voyager-verify` directory,
     // we'll read the files from there.
@@ -91,7 +87,7 @@ pub fn resolve_scarb(
 
     // The compiler compiles into the original scarb package name
     // As such we have to craft the correct path to the main package
-    let project_dir_path = extracted_files_dir.join(scarb_metadata_package_name);
+    let project_dir_path = extracted_files_dir.join(extracted_scarb_toml_data.name.clone());
     let project_dir_path = project_dir_path
         .strip_prefix(extracted_files_dir.clone())
         .unwrap();
@@ -102,13 +98,22 @@ pub fn resolve_scarb(
         .filter_map(|f| f.ok())
         .filter(|f| f.file_type().is_file())
         .filter(|f| {
-            f.path().extension().unwrap() == "cairo"
-                || f.path()
-                    .file_name()
-                    .map(|f| f.to_string_lossy().into_owned())
-                    .unwrap_or("".into())
-                    .to_lowercase()
-                    == "scarb.toml"
+            let file_path = f.path();
+
+            let is_cairo_file = match file_path.extension() {
+                Some(ext) => ext == "cairo",
+                None => false,
+            };
+            let file_entry_name = file_path
+                .file_name()
+                .map(|f| f.to_string_lossy().into_owned())
+                .unwrap_or("".into());
+
+            let is_supplementary_file = file_entry_name.to_lowercase() == "scarb.toml"
+                || file_entry_name == extracted_scarb_toml_data.license_file
+                || file_entry_name == extracted_scarb_toml_data.readme;
+
+            is_cairo_file || is_supplementary_file
         })
         .collect::<Vec<DirEntry>>();
 
@@ -130,14 +135,18 @@ pub fn resolve_scarb(
         })
         .collect::<Vec<FileInfo>>();
 
+    let contract_file = format!(
+        "{}/src/{}",
+        extracted_scarb_toml_data.name.clone(),
+        contract_paths[0].as_str()
+    );
+
     let project_metadata = ProjectMetadataInfo {
         cairo_version,
         scarb_version,
+        contract_file,
         project_dir_path: project_dir_path.as_str().to_owned(),
     };
-
-    pb.inc(1);
-    pb.finish_and_clear();
 
     Ok((project_files, project_metadata))
 }
