+++ crates/cli/src/verify.rs
@@ -2,14 +2,11 @@ use std::{env::current_dir, fs, str::FromStr, time::Instant};
 
 use anyhow::Result;
 use camino::Utf8PathBuf;
-use clap::{arg, builder::PossibleValue, Args, ValueEnum};
-use console::{style, Emoji};
-use indicatif::{HumanDuration, ProgressBar, ProgressStyle};
-use serde::{Deserialize, Serialize};
-use voyager_resolver_cairo::compiler::scarb_utils::read_additional_scarb_manifest_metadata;
-use walkdir::{DirEntry, WalkDir};
+use clap::{arg, Args};
+use console::Emoji;
+use indicatif::{HumanDuration, ProgressStyle};
 
-use dyn_compiler::dyn_compiler::{SupportedCairoVersions, SupportedScarbVersions};
+use dyn_compiler::dyn_compiler::SupportedCairoVersions;
 
 use crate::{
     api::{
@@ -58,8 +55,8 @@ pub struct VerifyFileArgs {
 
 pub fn verify_project(
     args: VerifyProjectArgs,
-    cairo_version: SupportedCairoVersions,
-    scarb_version: SupportedScarbVersions,
+    metadata: ProjectMetadataInfo,
+    files: Vec<FileInfo>,
 ) -> Result<()> {
     // Start a spinner for the verification process
     let started = Instant::now();
@@ -67,39 +64,12 @@ pub fn verify_project(
         .unwrap()
         .tick_chars("⠁⠂⠄⡀⢀⠠⠐⠈");
 
-    println!(
-        "{} {}Extracting files from the Scarb project...",
-        style("[1/4]").bold().dim(),
-        Emoji("📃  ", "")
-    );
-    // Extract necessary files from the Scarb project for the verified contract
-    let source_dir = if args.path.is_absolute() {
-        args.path
-    } else {
-        let mut current_path = current_dir().unwrap();
-        current_path.push(args.path);
-        Utf8PathBuf::from_path_buf(current_path).unwrap()
-    };
+    // println!(
+    //     "{} {}Checking if the class is already declared...",
+    //     style("[2/4]").bold().dim(),
+    //     Emoji("🔍  ", "")
+    // );
 
-    let compiler = get_dynamic_compiler(cairo_version);
-
-    let contract_paths = compiler.get_contracts_to_verify_path(&source_dir)?;
-
-    // TODO: maybe support multiple contracts in one verification?
-    if contract_paths.is_empty() {
-        return Err(anyhow::anyhow!("No contracts to verify"));
-    }
-    if contract_paths.len() > 1 {
-        return Err(anyhow::anyhow!(
-            "Only one contract can be verified at a time"
-        ));
-    }
-
-    println!(
-        "{} {}Checking if the class is already declared...",
-        style("[2/4]").bold().dim(),
-        Emoji("🔍  ", "")
-    );
     // Check if the class exists on the network
     let network_enum = Network::from_str(args.network.as_str())?;
     match does_class_exist(network_enum.clone(), &args.hash) {
@@ -113,99 +83,6 @@ pub fn verify_project(
         }
     }
 
-    println!(
-        "{} {}Resolving contract dependencies...",
-        style("[3/4]").bold().dim(),
-        Emoji("🔗  ", "")
-    );
-    let steps = 4;
-    let pb = ProgressBar::new(steps);
-
-    // Read the scarb metadata to get more information
-    // TODO: switch this to using scarb-metadata
-    let scarb_toml_content = fs::read_to_string(source_dir.join("Scarb.toml"))?;
-    let extracted_scarb_toml_data =
-        read_additional_scarb_manifest_metadata(scarb_toml_content.as_str())?;
-
-    // Compiler and extract the necessary files
-    compiler.compile_project(&source_dir)?;
-    pb.inc(2);
-
-    // Since we know that we extract the files into the `voyager-verify` directory,
-    // we'll read the files from there.
-    let extracted_files_dir = source_dir.join("voyager-verify");
-
-    // The compiler compiles into the original scarb package name
-    // As such we have to craft the correct path to the main package
-    let project_dir_path = extracted_files_dir.join(extracted_scarb_toml_data.name.clone());
-    let project_dir_path = project_dir_path
-        .strip_prefix(extracted_files_dir.clone())
-        .unwrap();
-
-    // Read project directory
-    let project_files = WalkDir::new(extracted_files_dir.as_path())
-        .into_iter()
-        .filter_map(|f| f.ok())
-        .filter(|f| f.file_type().is_file())
-        .filter(|f| {
-            let file_path = f.path();
-
-            let is_cairo_file = match file_path.extension() {
-                Some(ext) => ext == "cairo",
-                None => false,
-            };
-            let file_entry_name = file_path
-                .file_name()
-                .map(|f| f.to_string_lossy().into_owned())
-                .unwrap_or("".into());
-
-            let is_supplementary_file = file_entry_name.to_lowercase() == "scarb.toml"
-                || file_entry_name == extracted_scarb_toml_data.license_file
-                || file_entry_name == extracted_scarb_toml_data.readme;
-
-            is_cairo_file || is_supplementary_file
-        })
-        .collect::<Vec<DirEntry>>();
-
-    let project_files = project_files
-        .iter()
-        .map(|f| {
-            let actual_path = f.path().to_owned();
-            let file_name = actual_path
-                .strip_prefix(&extracted_files_dir)
-                .unwrap()
-                .to_str()
-                .to_owned()
-                .unwrap()
-                .to_string();
-            FileInfo {
-                name: file_name,
-                path: actual_path,
-            }
-        })
-        .collect::<Vec<FileInfo>>();
-    pb.inc(1);
-    pb.finish_and_clear();
-
-    // We already know the contract file specified in Scarb.toml is relative to src/
-    let contract_file = format!(
-        "{}/src/{}",
-        extracted_scarb_toml_data.name.clone(),
-        contract_paths[0].as_str()
-    );
-    // let spinner = ProgressBar::new_spinner();
-    // spinner.set_style(ProgressStyle::default_spinner());
-    // spinner.set_style(ProgressStyle::with_template("{spinner:2.green/white} {msg} [{elapsed_precise}] ").unwrap());
-    // spinner.set_message("dispatching verification job");
-
-    // let spinner_clone = spinner.clone();
-    // thread::spawn(move || {
-    //     while !spinner_clone.is_finished() {
-    //         spinner_clone.tick();
-    //         thread::sleep(std::time::Duration::from_millis(100));
-    //     }
-    // });
-
     let dispatch_response = dispatch_class_verification_job(
         args.api_key.as_str(),
         network_enum.clone(),
@@ -213,13 +90,8 @@ pub fn verify_project(
         args.license.to_long_string().as_str(),
         args.is_account_contract.unwrap_or(false),
         &args.name,
-        ProjectMetadataInfo {
-            cairo_version,
-            scarb_version,
-            contract_file,
-            project_dir_path: project_dir_path.as_str().to_owned(),
-        },
-        project_files,
+        metadata,
+        files,
     );
 
     let job_id = match dispatch_response {
